fix: audit corrections — validation, data integrity, error handling
- Products: fix invalid Prisma API usage (product.fields.minStock) with in-memory filter
- Customers: redeemPoints validates balance before decrementing, proper error messages
- Stock: adjustInventory rejects negative stock values
- Shifts: closeShift validates shift exists and is open before closing
- Shifts: Z-report fallback only when registerId is null (avoid cross-register contamination)
- Import: parseFloat for price fields (was parseInt, losing decimals)
- Transactions: attach _warnings object with stockErrors/loyaltyError for frontend visibility

+++ src/import/import.service.ts
@@ -99,8 +99,8 @@ export class ImportService {
             throw new BadRequestException('name est obligatoire');
           }
 
-          const price = parseInt(data.price) || 0;
-          const costPrice = parseInt(data.costprice || data.cost_price || '0') || 0;
+          const price = parseFloat(data.price) || 0;
+          const costPrice = parseFloat(data.costprice || data.cost_price || '0') || 0;
           const stock = parseInt(data.stock) || 0;
           const minStock = parseInt(data.minstock || data.min_stock || '10') || 10;
